+++ jexxa-core/src/test/java/io/jexxa/infrastructure/drivenadapterstrategy/persistence/jdbc/JDBCTestDatabase.java
@@ -24,16 +24,17 @@ import java.util.stream.Stream;
 
 public class JDBCTestDatabase
 {
+    static final String ADMIN = "admin";
     static final int PRIMARY_KEY_WITH_NULL_VALUES = 1;
     static final int PRIMARY_KEY_WITH_NONNULL_VALUES = 2;
     static final int PRIMARY_KEY_WITH_NONNULL_VALUES_DUPLICATE = 3;
     static final int PRIMARY_KEY_VALUES_NOT_PRESENT = 4;
-    static final Timestamp testTimestamp = Timestamp.from(Instant.now().truncatedTo(ChronoUnit.MICROS));
-    static final String testString = "Hello World";
-    static final int testIntValue = 2;
-    static final int testFloatValue = 3;
-    static final int testDoubleValue = 4;
-    static final BigDecimal testNumericValue = BigDecimal.valueOf(5);
+    static final Timestamp TEST_TIMESTAMP = Timestamp.from(Instant.now().truncatedTo(ChronoUnit.MICROS));
+    static final String TEST_STRING = "Hello World";
+    static final int TEST_INT_VALUE = 2;
+    static final int TEST_FLOAT_VALUE = 3;
+    static final int TEST_DOUBLE_VALUE = 4;
+    static final BigDecimal TEST_NUMERIC_VALUE = BigDecimal.valueOf(5);
 
 
     enum JDBCTestSchema
@@ -53,20 +54,25 @@ public class JDBCTestDatabase
     static Stream<Properties> repositoryConfig() {
         var postgresProperties = new Properties();
         postgresProperties.put(JDBCConnection.JDBC_DRIVER, "org.postgresql.Driver");
-        postgresProperties.put(JDBCConnection.JDBC_PASSWORD, "admin");
-        postgresProperties.put(JDBCConnection.JDBC_USERNAME, "admin");
+        postgresProperties.put(JDBCConnection.JDBC_PASSWORD, ADMIN);
+        postgresProperties.put(JDBCConnection.JDBC_USERNAME, ADMIN);
         postgresProperties.put(JDBCConnection.JDBC_URL, "jdbc:postgresql://localhost:5432/jexxa");
         postgresProperties.put(JDBCConnection.JDBC_AUTOCREATE_TABLE, "true");
         postgresProperties.put(JDBCConnection.JDBC_AUTOCREATE_DATABASE, "jdbc:postgresql://localhost:5432/postgres");
 
         var h2Properties = new Properties();
         h2Properties.put(JDBCConnection.JDBC_DRIVER, "org.h2.Driver");
-        h2Properties.put(JDBCConnection.JDBC_PASSWORD, "admin");
-        h2Properties.put(JDBCConnection.JDBC_USERNAME, "admin");
+        h2Properties.put(JDBCConnection.JDBC_PASSWORD, ADMIN);
+        h2Properties.put(JDBCConnection.JDBC_USERNAME, ADMIN);
         h2Properties.put(JDBCConnection.JDBC_URL, "jdbc:h2:mem:jexxa;DB_CLOSE_DELAY=-1");
         h2Properties.put(JDBCConnection.JDBC_AUTOCREATE_TABLE, "true");
 
         return Stream.of(postgresProperties, h2Properties);
+    }
+
+    public static Properties getPostgresProperties()
+    {
+        return repositoryConfig().findFirst().orElseThrow();
     }
 
 
@@ -114,12 +120,13 @@ public class JDBCTestDatabase
 
         var insertNonNullValues = jdbcConnection.createCommand(JDBCTestSchema.class)
                 .insertInto(JDBCTestDatabase.class)
-                .values(PRIMARY_KEY_WITH_NONNULL_VALUES, testIntValue , testNumericValue, testFloatValue, testDoubleValue, testString, testTimestamp)
+                .values(PRIMARY_KEY_WITH_NONNULL_VALUES, TEST_INT_VALUE, TEST_NUMERIC_VALUE, TEST_FLOAT_VALUE, TEST_DOUBLE_VALUE, TEST_STRING, TEST_TIMESTAMP)
                 .create();
 
         var insertNonNullValuesDuplicate = jdbcConnection.createCommand(JDBCTestSchema.class)
                 .insertInto(JDBCTestDatabase.class)
-                .values(PRIMARY_KEY_WITH_NONNULL_VALUES_DUPLICATE, testIntValue , testNumericValue, testFloatValue, testDoubleValue, testString, testTimestamp)
+                .values(PRIMARY_KEY_WITH_NONNULL_VALUES_DUPLICATE, TEST_INT_VALUE, TEST_NUMERIC_VALUE, TEST_FLOAT_VALUE, TEST_DOUBLE_VALUE, TEST_STRING,
+                        TEST_TIMESTAMP)
                 .create();
 
         insertNullValues.asUpdate();