+++ jexxa-web/src/main/java/io/jexxa/infrastructure/drivingadapter/rest/openapi/OpenAPIConvention.java
@@ -19,14 +19,16 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Properties;
 
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.util.StdDateFormat;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.google.gson.JsonParser;
+import com.kjetland.jackson.jsonSchema.JsonSchemaGenerator;
 import io.javalin.core.JavalinConfig;
 import io.javalin.plugin.openapi.OpenApiOptions;
 import io.javalin.plugin.openapi.OpenApiPlugin;
@@ -35,6 +37,7 @@ import io.javalin.plugin.openapi.dsl.DocumentedContent;
 import io.javalin.plugin.openapi.dsl.OpenApiBuilder;
 import io.javalin.plugin.openapi.dsl.OpenApiDocumentation;
 import io.jexxa.utils.JexxaLogger;
+import io.jexxa.utils.json.JSONManager;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.media.ArraySchema;
 import io.swagger.v3.oas.models.media.BooleanSchema;
@@ -55,13 +58,11 @@ public class OpenAPIConvention
     private final Properties properties;
     private final JavalinConfig javalinConfig;
     private OpenApiOptions openApiOptions;
-    private final GsonBuilder gsonBuilder;
 
-    public OpenAPIConvention(Properties properties, JavalinConfig javalinConfig, GsonBuilder gsonBuilder)
+    public OpenAPIConvention(Properties properties, JavalinConfig javalinConfig)
     {
         this.properties = properties;
         this.javalinConfig = javalinConfig;
-        this.gsonBuilder = gsonBuilder;
         initOpenAPI();
     }
     private void initOpenAPI()
@@ -75,6 +76,9 @@ public class OpenAPIConvention
 
             openApiOptions = new OpenApiOptions(applicationInfo)
                     .path("/" + properties.getProperty(OPEN_API_PATH));
+
+            //Show all fields of an ValueObject
+            openApiOptions.getJacksonMapper().setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
 
             javalinConfig.registerPlugin(new OpenApiPlugin(openApiOptions));
             javalinConfig.enableCorsForAllOrigins();
@@ -131,7 +135,7 @@ public class OpenAPIConvention
                     openApiOperation.operationId(method.getName());
                 });
 
-        documentParameters(method, openApiDocumentation, gsonBuilder);
+        documentParameters(method, openApiDocumentation);
 
         documentReturnType(method, openApiDocumentation);
 
@@ -152,12 +156,12 @@ public class OpenAPIConvention
         }
     }
 
-    private static void documentParameters(Method method, OpenApiDocumentation openApiDocumentation, GsonBuilder gsonBuilder)
+    private static void documentParameters(Method method, OpenApiDocumentation openApiDocumentation)
     {
         if (method.getParameters().length == 1 )
         {
             var schema = createSchema(method.getParameterTypes()[0], method.getGenericParameterTypes()[0]);
-            schema.setExample(createExampleInstance(method.getParameterTypes()[0], method.getGenericParameterTypes()[0], gsonBuilder));
+            schema.setExample(createExampleInstance(method.getParameterTypes()[0], method.getGenericParameterTypes()[0]));
 
             //For some reason I have to add requests as DocumentedContent. Otherwise components seems to be not correctly created
             openApiDocumentation.body(List.of(new DocumentedContent(method.getParameterTypes()[0], isJsonArray(method.getParameterTypes()[0]), APPLICATION_TYPE_JSON )));
@@ -170,7 +174,7 @@ public class OpenAPIConvention
 
             for (int i = 0; i < method.getParameterTypes().length; ++i)
             {
-                exampleObjects[i] = createExampleInstance(method.getParameterTypes()[i],method.getGenericParameterTypes()[i], gsonBuilder);
+                exampleObjects[i] = createExampleInstance(method.getParameterTypes()[i],method.getGenericParameterTypes()[i]);
                 var parameterSchema = createSchema(method.getParameterTypes()[i], method.getGenericParameterTypes()[i]);
                 parameterSchema.setExample(exampleObjects[i]);
 
@@ -195,20 +199,22 @@ public class OpenAPIConvention
         mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
         // StdDateFormat is ISO8601 since jackson 2.9
         mapper.setDateFormat(new StdDateFormat().withColonInTimeZone(true));
+        mapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
 
-        var schema = mapper.generateJsonSchema(clazz);
+        JsonSchemaGenerator jsonSchemaGenerator = new JsonSchemaGenerator(mapper);
+        JsonNode jsonSchema = jsonSchemaGenerator.generateJsonSchema(clazz);
 
-        return schema.toString();
+        return mapper.writeValueAsString(jsonSchema);
     }
 
-    private static Object createExampleInstance(Class<?> clazz, Type genericType, GsonBuilder gsonBuilder)
+    private static Object createExampleInstance(Class<?> clazz, Type genericType)
     {
         if ( isJava8Date(clazz) )
         {
             return java8DateExample(clazz);
         }
 
-        return createGenericExample(clazz, genericType, gsonBuilder);
+        return createGenericExample(clazz, genericType);
     }
 
     private static Object java8DateExample(Class<?> clazz)
@@ -231,7 +237,7 @@ public class OpenAPIConvention
         return null;
     }
 
-    private static Object createGenericExample(Class<?> clazz, Type genericType, GsonBuilder gsonBuilder)
+    private static Object createGenericExample(Class<?> clazz, Type genericType)
     {
         try
         {
@@ -256,15 +262,14 @@ public class OpenAPIConvention
                     return null;
                 }
 
-                Gson gson = gsonBuilder.create();
-                return gson.fromJson(jsonObject, clazz);
+                return JSONManager.getJSONConverter().fromJson(jsonObject.toString(), clazz);
             }
 
             //Handle JsonArray
             if (typeInformation.getAsString().equals(JSON_ARRAY_TYPE))
             {
                 var result = new Object[1];
-                result[0] = createExampleInstance(extractTypeFromArray(genericType), null, gsonBuilder);
+                result[0] = createExampleInstance(extractTypeFromArray(genericType), null);
                 return result;
             }
 
@@ -386,7 +391,7 @@ public class OpenAPIConvention
         return (Class<?>)parameterType.getActualTypeArguments()[0];
     }
 
-    @SuppressWarnings({"java:S1104", "java:S116"})
+    @SuppressWarnings({"java:S1104", "java:S116", "unused"})
     public static class BadRequestResponse
     {
         public String Exception;