+++ jexxa-web/src/test/java/io/jexxa/infrastructure/drivingadapter/rest/HttpsRESTfulRPCAdapterIT.java
@@ -1,5 +1,7 @@
 package io.jexxa.infrastructure.drivingadapter.rest;
 
+import static io.jexxa.infrastructure.drivingadapter.rest.RESTConstants.APPLICATION_TYPE;
+import static io.jexxa.infrastructure.drivingadapter.rest.RESTConstants.CONTENT_TYPE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
@@ -22,14 +24,13 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.ssl.SSLContextBuilder;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class HttpsRESTfulRPCAdapterIT
 {
-    private static final String CONTENT_TYPE = "Content-Type";
-    private static final String APPLICATION_TYPE = "application/json";
     private static final String METHOD_GET_SIMPLE_VALUE = "getSimpleValue";
 
     private static final int DEFAULT_VALUE = 42;
@@ -88,8 +89,8 @@ class HttpsRESTfulRPCAdapterIT
 
         //Assert
         assertNotNull(result);
-        assertEquals(DEFAULT_VALUE, simpleApplicationService.getSimpleValue());
-        assertEquals(simpleApplicationService.getSimpleValue(), result.intValue() );
+        Assertions.assertEquals(DEFAULT_VALUE, simpleApplicationService.getSimpleValue());
+        Assertions.assertEquals(simpleApplicationService.getSimpleValue(), result.intValue() );
     }
 
     @AfterEach