+++ jexxa-web/src/test/java/io/jexxa/infrastructure/drivingadapter/rest/OpenAPIJava8DateTimeTest.java
@@ -1,5 +1,7 @@
 package io.jexxa.infrastructure.drivingadapter.rest;
 
+import static io.jexxa.infrastructure.drivingadapter.rest.RESTConstants.APPLICATION_TYPE;
+import static io.jexxa.infrastructure.drivingadapter.rest.RESTConstants.CONTENT_TYPE;
 import static io.jexxa.infrastructure.drivingadapter.rest.RESTfulRPCAdapter.HTTP_PORT_PROPERTY;
 import static io.jexxa.infrastructure.drivingadapter.rest.RESTfulRPCAdapter.OPEN_API_PATH;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -14,7 +16,7 @@ import java.util.Properties;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import io.jexxa.TestConstants;
-import io.jexxa.application.applicationservice.SimpleApplicationService;
+import io.jexxa.application.applicationservice.Java8DateTimeApplicationService;
 import kong.unirest.Unirest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -25,13 +27,10 @@ import org.junit.jupiter.api.parallel.ExecutionMode;
 
 @Execution(ExecutionMode.SAME_THREAD)
 @Tag(TestConstants.INTEGRATION_TEST)
-class OpenAPITest
+class OpenAPIJava8DateTimeTest
 {
-    private static final String CONTENT_TYPE = "Content-Type";
-    private static final String APPLICATION_TYPE = "application/json";
-
     private static final String OPENAPI_PATH = "http://localhost:7000/swagger-docs/";
-    private final SimpleApplicationService simpleApplicationService = new SimpleApplicationService();
+    private final Java8DateTimeApplicationService java8DateTimeApplicationService = new Java8DateTimeApplicationService();
 
 
     private RESTfulRPCAdapter objectUnderTest;
@@ -48,7 +47,7 @@ class OpenAPITest
         properties.put(OPEN_API_PATH, "swagger-docs");
 
         objectUnderTest = RESTfulRPCAdapter.createAdapter(properties);
-        objectUnderTest.register(simpleApplicationService);
+        objectUnderTest.register(java8DateTimeApplicationService);
         objectUnderTest.start();
     }
 
@@ -67,7 +66,7 @@ class OpenAPITest
         //Arrange -> Nothing to do
 
         //Act
-        JsonObject result = Unirest.get(OPENAPI_PATH)
+       JsonObject result = Unirest.get(OPENAPI_PATH)
                 .header(CONTENT_TYPE, APPLICATION_TYPE)
                 .asObject(JsonObject.class).getBody();
 
@@ -95,11 +94,11 @@ class OpenAPITest
         result.forEach( element -> assertNotNull( element.getAsJsonObject().get(APPLICATION_TYPE) ) );
     }
 
-    @Test
+   @Test
     void testHTTPRequestMapping()
     {
         //Arrange
-        var resTfulRPCConvention = new RESTfulRPCConvention(simpleApplicationService);
+        var resTfulRPCConvention = new RESTfulRPCConvention(java8DateTimeApplicationService);
 
         JsonObject openAPI = Unirest.get(OPENAPI_PATH)
                 .header(CONTENT_TYPE, APPLICATION_TYPE)