+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## \[3.0.0] - yyyy-mm-dd
+### Changed
+-   Introduced new package Jexxa-Web which includes all web specific plugins such as `RESTfulRPCAdapter`. See [reference guide](https://repplix.github.io/Jexxa/jexxa_reference.html#_jexxa_modules) for more information. If your application uses this adapter, you have now to add dependency `jexxa-web`. See tutorial [HelloJexxa](https://github.com/repplix/Jexxa/blob/master/tutorials/HelloJexxa/README.md) for an example.  
+-   Removed all methods and constants that were declared deprecated in 2.x
+
+### Fixed
+-   `RESTfulRPCAdapter`: 
+    -   OpenAPI plugin uses global JSONManager for creating prototypes so that specific type converter can be registered    
+    
 ## \[2.8.1] - 2021-02-27
 ### Fixed
 -   Updated JaCoCo plugin so that tests run under Java 15